moved the brand, hotel, and floor images from the website into the ARM folder to be used during the provisioning.
updated the provisioning generator to now output a BlobDescription on SpaceDescriptions for Brand, Hotel, and Floor instead of the ImagePath property. It is now pointing to the image file (relatively to the template file) that needs to be uploaded to blob storage.

provisioning will be updated to upload the file and then store the image blob's id as a custom property instead of ImagePath